"Removed .eslintrc.cjs file, updated dependencies in frontend and server package.json files, and modified DeviceComputerSchema.ts file"

+++ server/src/Contexts/Features/Computer/infraestructure/sequelize/DeviceComputerSchema.ts
@@ -78,7 +78,7 @@ export function initDeviceComputerModel (sequelize: Sequelize): void {
         allowNull: true
       },
       memoryRam: {
-        type: DataTypes.ARRAY(DataTypes.INTEGER),
+        type: DataTypes.ARRAY(DataTypes.DECIMAL),
         allowNull: true
       },
       memoryRamCapacity: {
